Create GetCerts.cs
F5 will give you a decent report of all your certificates and their expiration dates. However it does not tell you what Virtual Server or SSL Profile the certificates are applied to (or if they are used at all). What this code does is grabs all the virtual servers, all SSL profiles, and all certificates. Then it loops through them to find where a certificate is applied. Then it returns the certificate and virtual server info. I wrote it in C# but the logic can be used anywhere and the API calls are independent of language.

Requirements
API Access, Newtonsoft.Json (for C# add a reference to it via the Solution Explorer > Manage NuGet Packages)